[FEAT] 뱃지추가 구현

+++ src/main/java/com/d204/rumeet/badge/model/dto/MyBadgeDto.java
@@ -9,5 +9,5 @@ import lombok.Data;
 public class MyBadgeDto {
     int userId;
     int badgeId;
-    int date;
+    long date;
 }
